pyproject.toml
[tool.poetry]
name = "app"
version = "0.1.0"
description = ""
authors = ["shk"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
weaviate-client = "^4.7.0"
pydantic-settings = "^2.3.4"
openai = "^1.37.0"
minio = "^7.2.7"


[tool.poetry.group.dev.dependencies]
black = "^24.4.2"
pytest = "^8.3.1"
pytest-cov = "^5.0.0"
isort = "^5.13.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
